+++ CHANGES.md
@@ -8,6 +8,7 @@ Handy Git hooks to integrate with [pre-commit](http://pre-commit.com/) framework
 
 * v1.2.0 (2021-09-24)
   * `checkstyle-jar` now can automatically download and cache JAR file.
+  * `checkstyle-jar` accepts `--config` argument now.
   * `end-of-file` and `trailing-whitespaces`'s `--fix` now requires "yes/no" argument.
 
 * v1.1.0 (2021-09-23)